WIP: Add support for Upvalues
Upvalues store the captured variables in closures.
We look for the local variable in the enclosing function and store a
reference to it.
If not found, the variable must be global.

However, to handle deeply nested functions, resolve_upvalue()
is a recursive function. This allows it to store a ref to either a
local or another upvalue.