Track which locals are captured by a closure
In resolve_upvalue(), if we grab a local, we mark it so that it can be
treated seperately to the locals not captured.

Upvalues have to be closed when the local they refer to are not on the
stack anymore. This means, we use a flag. When ending a new scope, we
pop the non-captured locals, and emit OP_CLOSE_UPVALUE for those which
have to be closed.