Expose using 'this' to access the instance
When we get a 'this' token, we just declare it as a local variable which
cannot be assigned to.
Adding a new FuncType so that it can tell if it is parsing a method or
not.

Slot zero for a CallFrame is repurposed so that we can put the receiver
of the bound method in it.
When it looks for 'this' it just finds that slot and uses it.