Call init() when creating an instance
When an instance is being created, we look for 'init' method and invoke
it, if found. Otherwise, we raise an error if there are args being
passed.

TODO: Make init() implicitly return the instance.
Right now, it returns nil, so users will have to explicitly 'return
this;' at the end.