Emit OP_INVOKE if we see a direct method call
Before, we used to push the instance,, then pop it off and push the
ObjBoundMethod and pop it off for ObjClosure and then call it.

This is a very slow way of doing things, so if we see a direct method
call, we emit a different opcode to do it all at once.